Windows, build swig interfac oc_connectivity.i on debug
Due to a configuration issue the build was only building
oc_connectivity.i when the release build option was selected

This change will cause oc_connectivity.i to be built for both
release and debug build.

Additionally oc_connectivity_wrap.c was also added to the build
of the iotivity-lite-jni project with out adding that none of
the OCConnectivity functions can be called from Java without
causing a library link error.

Signed-off-by: George Nash <george.nash@intel.com>