Update introspection data for smart home server
Add some startup printf statments to The smart_home_server_linux
When running a release build the lack of any print statments
made it really hard to tell if the sample was running as expected.

Updated the max app data size to 18K bytes to fit the introspection
data.

Added some print statments for the common errors encontered when
onboarding the introspectionclient

Signed-off-by: George Nash <george.nash@intel.com>