Add clang-format tool to the CI build
This adds the clang-format tool to the CI build. This enables
catching formating errors on from contributers that have not
installed the pre-commit hook.

This change was integrated into the same build as the doxygen
build script.  This was done to reduce the number of CI build
pipelines needed to pass the build.

Copyright block added to the build_doc.sh script. The scripts
also print out some blocks of text to indicate there run status
making it easer to instantly see the results of the build
output.

The new script added to the .gitignore if it is found in the
top of the project  incase users want to use it localy on
their own computers. Similar to the reason the _clang-format
file is in the ignore file.

Signed-off-by: George Nash <george.nash@intel.com>